refactor: 패키지 세분화 및 LottoController에서 static 제거
