+++ src/test/java/lotto/domain/lotto/LottosTest.java
@@ -1,5 +1,6 @@
 package lotto.domain.lotto;
 
+import lotto.domain.lottostrategy.LottoFactory;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EmptySource;
